✨ 引入在线 TTS 的 Token 管理与自动刷新机制
- **新增 Token 管理核心 (`WxTokenManager`, `TokenProvider`)**：
  - 新增 `WxTokenManager` 作为全局单例入口，负责初始化和提供 `TokenProvider`。
  - `TokenProvider` 实现了 Token 的内存与 SharedPreferences 缓存、并发安全的“单飞”刷新逻辑，避免多个请求同时刷新 Token。
  - 新增 `TokenRemoteDataSource`，用于从内网服务 (`http://192.168.1.212:8866`) 获取最新的有效 Token。

- **实现 Token 自动刷新与重试**：
  - 在 `WxReaderApi` 中集成 `TokenProvider`，当 TTS 请求因 Token 过期（返回 -13）失败时，能自动触发一次 Token 刷新并重试请求。
  - `MainActivity` 在启动时会调用 `WxTokenManager.refreshTokenIfNeed`，在 Token 不存在或超过 24 小时的情况下主动刷新，以提高首次合成成功率。

- **API 与配置变更**：
  - 将 `uid` 的类型从 `Int` 统一修改为 `Long`，以支持更大的用户 ID 范围，涉及 `OnlineTtsApi`、`TtsSynthesizer` 等多个接口。（**破坏性变更**）
  - 在 `AndroidManifest.xml` 中配置 `networkSecurityConfig`，允许应用在 Debug 和 Release 版本中对内网 Token 服务 IP (`192.168.1.212`) 发起明文 HTTP 请求。

- **日志与代码优化**：
  - 引入 `LogMask` 工具类对日志中的 `token` 和 `uid` 进行脱敏处理，增强安全性。
  - 隐藏了 `MainActivity` 界面上用于手动输入 Token/UID 的输入框，改为依赖自动刷新机制。

+++ TTSLibrary/src/main/java/com/qq/wx/offlinevoice/synthesizer/TtsSynthesizer.kt
@@ -322,7 +322,7 @@ class TtsSynthesizer(
      */
     fun seekToSentence(index: Int) = sendCommand(Command.SeekTo(index)) // index 按“逻辑行”语义
 
-    fun setToken(token: String, uid: Int) {
+    fun setToken(token: String, uid: Long) {
         ttsRepository.onlineApi.setToken(token, uid)
     }
